Split Select component from input.js
and fix eslint and flow errors.
