+++ web/html/src/manager/image-import.js
@@ -6,7 +6,8 @@ const React = require("react");
 const ReactDOM = require("react-dom");
 const { TopPanel } = require('components/panels/TopPanel');
 const { Text } = require('components/input/Text');
-const {Form, Select} = require("../components/input");
+const { Select } = require('components/input/Select');
+const { Form } = require('components/input');
 const {SubmitButton, Button} = require("../components/buttons");
 const Network = require("../utils/network");
 const {Messages} = require("../components/messages");
