+++ web/html/src/manager/image-store-edit.js
@@ -10,6 +10,7 @@ const {SubmitButton, Button} = require("../components/buttons");
 const Input = require("../components/input");
 const { Check } = require('components/input/Check');
 const { Password } = require('components/input/Password');
+const { Select } = require('components/input/Select');
 const { Text } = require('components/input/Text');
 const Utils = require("../utils/functions").Utils;
 
@@ -175,13 +176,13 @@ class CreateImageStore extends React.Component {
           onChange={this.onFormChange}
           onSubmit={(e) => this.isEdit() ? this.onUpdate(e) : this.onCreate(e)}
           onValidate={this.onValidate}>
-          <Input.Select labelClass="col-md-3" divClass="col-md-6" label={t("Store Type")} name="storeType" required disabled={this.isEdit()}>
+          <Select labelClass="col-md-3" divClass="col-md-6" label={t("Store Type")} name="storeType" required disabled={this.isEdit()}>
             {
               this.state.storeTypes.map(k =>
                 <option key={k} value={k}>{ typeMap[k] }</option>
               )
             }
-          </Input.Select>
+          </Select>
           <Text name="label" label={t("Label")} required validators={this.isLabelUnique} invalidHint={t("Label is required and must be unique.")} labelClass="col-md-3" divClass="col-md-6"/>
           <Text name="uri" label={t("Store URI")} required hint={<span>The URI to the store's API endpoint</span>} labelClass="col-md-3" divClass="col-md-6"/>
           { this.renderTypeInputs(this.state.model.storeType) }
